Fixed the Get_abstract for the json, i used the imp_sent_creator now it's easy to do it with the other one if needed but the result wont be as good. I modified the imp_sent_creator so it allows for a modifiable amount of sentences. I picked 3 sentences for the Json part since that was requested by Nicklas verbally during the meeting.
I also sent in a new function in the insert file. that handles multiple files where the names are in a array.
Talked about it with fabian like a week ago but there hasn't been an issue for it.
I decided to include it here since it's very small so making a own issue for it may be just a waste of time.